core: port to GtkApplication and use the new application menus
This patch evolves the `DhBase' into a fully featured new `DhApp', which
inherits from GtkApplication. The new `DhApp' will take care of every
`DhWindow' and `DhAssistant' associated to the application.

Application-specific actions are added to the new application menu:
 * New window
 * Preferences
 * About Devhelp
 * Quit

As part of this process of setting up the menu, the Preferences dialog
was modified not to be tied to any specific window.

https://bugzilla.gnome.org/show_bug.cgi?id=677927